Fix for UI.Text allign
align is used to both set the alignment in the front string and to physically move the location where the font string is to be. The end result is that align is always centered.

This fixes this by making the default alignment centred and removing the physical positioning of the text.Bounding points are adjusted accordingly.