Added a volume control
The Audio object now has a volume property that can be set. This uses a
GainNode which connects the sound asset to the destination. I think it
makes sense to have this control on the same settings level as channels.
This doesn't work with IE11 since it still doesn't support the Web Audio
API.